+++ RayTracingInOneWeekend/Hittables/Hittable.cs
@@ -2,8 +2,8 @@
 {
     public abstract class Hittable
     {
-        public virtual bool Hit(Ray ray, double tMin, double tMax, ref HitRecord hitRecord) => false;
+        public virtual bool Hit(Ray ray, float tMin, float tMax, ref HitRecord hitRecord) => false;
 
-        public abstract bool BoundingBox(double time0, double time1, out AxisAlignedBoundingBox? boundingBox);
+        public abstract bool BoundingBox(float time0, float time1, out AxisAlignedBoundingBox? boundingBox);
     }
 }
